Attempt to set up microbundle
