+++ docs/gatsby-config.js
@@ -12,7 +12,7 @@ module.exports = {
       resolve: 'gatsby-plugin-compile-es6-packages',
       options: {
         modules: [
-          'theme-ui',
+          // 'theme-ui',
         ]
       }
     },
