+++ theme-ui/.gitignore
@@ -1,1 +1,2 @@
 coverage
+dist
