+++ theme-ui/.npmignore
@@ -2,3 +2,4 @@ babel.config.js
 *.test.js
 coverage
 __snapshots__
+test
