+++ theme-ui/package.json
@@ -1,11 +1,13 @@
 {
   "name": "theme-ui",
   "version": "0.0.8",
-  "main": "index.js",
+  "main": "dist/index.js",
+  "module": "dist/index.js",
   "author": "Brent Jackson <jxnblk@gmail.com>",
   "license": "MIT",
   "repository": "system-ui/theme-ui",
   "scripts": {
+    "prepare": "microbundle",
     "test": "jest"
   },
   "dependencies": {
@@ -14,6 +16,7 @@
     "jest-emotion": "^10.0.10",
     "lodash.get": "^4.4.2",
     "lodash.merge": "^4.6.1",
+    "microbundle": "^0.11.0",
     "styled-system": "^4.1.0",
     "typography": "^0.16.19"
   },
@@ -36,6 +39,10 @@
     "@emotion/styled": "^10.0.10",
     "@mdx-js/mdx": "^1.0.0-rc.0",
     "@mdx-js/react": "^1.0.0-rc.0",
-    "emotion-theming": "^10.0.10"
+    "emotion-theming": "^10.0.10",
+    "react": "^16.8.6"
+  },
+  "jest": {
+    "testRegex": "./test/.*.js$"
   }
 }
