+++ theme-ui/src/index.js
@@ -1,0 +1,21 @@
+export {
+  css,
+  Styled,
+  Context,
+  ComponentProvider,
+  // deprecate
+  useComponents,
+  ThemeProvider
+} from './core'
+
+export {
+  Box,
+  Flex,
+  Layout,
+  Header,
+  Main,
+  Container,
+  Footer,
+} from './layout'
+
+export { default as typography } from './typography'
