+++ theme-ui/test/typography.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import toStyles from './typography'
+import toStyles from '../src/typography'
 import theme from 'typography-theme-wordpress-2016'
 
 test('converts typography theme to styles object', () => {
